requirements.txt
altair==5.3.0
attrs==24.2.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
distro==1.9.0
et-xmlfile==1.1.0
Flask==3.0.3
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-api-python-client==2.138.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
httplib2==0.22.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jpype1==1.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==2.0.1
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pdfminer.six==20231228
pdfplumber==0.11.2
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.27.2
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
six==1.16.0
smmap==5.0.1
streamlit==1.37.1
tabula-py==2.10.0
tenacity==8.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
Werkzeug==3.0.3
xlrd==2.0.1
